+++ app/src/main/java/com/example/pooptube/home/HomeFragment.kt
@@ -6,6 +6,7 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.Toast
+import androidx.fragment.app.viewModels
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.example.pooptube.BuildConfig
 import com.example.pooptube.databinding.FragmentHomeBinding
@@ -27,13 +28,14 @@ class HomeFragment : Fragment() {
     private lateinit var binding: FragmentHomeBinding
     private lateinit var homeChipAdapter: HomeChipAdapter
     private lateinit var homeVideoAdapter: HomeVideoAdapter
-    private val apiService = ApiConfig.getService()
-//    private var selectedVideoData: VideosModelList? = null
+    private val viewModel : HomeViewModel by viewModels()
 
 
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
         binding = FragmentHomeBinding.inflate(inflater, container, false)
         return binding.root
+
+
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
@@ -63,7 +65,6 @@ class HomeFragment : Fragment() {
             }
         })*/
 
-
         with(binding) {
             chipRecyclerView.itemAnimator = null
             chipRecyclerView.adapter = homeChipAdapter
@@ -78,58 +79,25 @@ class HomeFragment : Fragment() {
             fetchPopularVideos()
         }
 
-        // Chip 카테고리ID 설정
-        val categoryFilterData = listOf(
-            HomeFilterModel(category = "음악", categoryId = "10"),
-            HomeFilterModel(category = "스포츠", categoryId = "17"),
-            HomeFilterModel(category = "게임", categoryId = "20"),
-            HomeFilterModel(category = "동물", categoryId = "15"),
-            HomeFilterModel(category = "엔터", categoryId = "26"),
-            HomeFilterModel(category = "테크", categoryId = "28"),
-            HomeFilterModel(category = "뉴스", categoryId = "25")
-        )
-        homeChipAdapter.setItems(categoryFilterData)
+        // ViewModel에서 categoryFilterData 관리
+        homeChipAdapter.setItems(viewModel.categoryFilterData)
+
+        viewModel.videos.observe(viewLifecycleOwner) { videoModels ->
+            binding.progressBar.visibility = View.GONE
+            homeVideoAdapter.setItems(videoModels)
+        }
+
+        viewModel.error.observe(viewLifecycleOwner) { errorMsg ->
+            binding.progressBar.visibility = View.GONE
+            Toast.makeText(requireContext(), errorMsg, Toast.LENGTH_SHORT).show()
+        }
+
+        fetchPopularVideos() // 처음에 기본 데이터 로드
     }
 
     private fun fetchPopularVideos(categoryId: String? = null) {
-        // 로딩 시작
         binding.progressBar.visibility = View.VISIBLE
-        val call = apiService.getVideoInfo(apiKey = BuildConfig.YOUTUBE_API_KEY, categoryId = categoryId, regionCode = "KR")
-
-        call.enqueue(object : Callback<VideosModelList> {
-            override fun onResponse(call: Call<VideosModelList>, response: Response<VideosModelList>) {
-                // 로딩 완료
-                binding.progressBar.visibility = View.GONE
-                if (response.isSuccessful) {
-                    val videoItems = response.body()?.items ?: emptyList()
-
-                    // YoutubeVideoItem 리스트를 HomeVideoModel 리스트로 변환
-                    val videoModels = videoItems.map { videoItem ->
-                        HomeVideoModel(
-                            imgThumbnail = videoItem.snippet.thumbnails.medium.url,
-                            channelLogoImg = "...", // 이부분은 추후에 적절한 값을 넣어야 합니다.
-                            title = videoItem.snippet.title,
-                            author = videoItem.snippet.channelTitle,
-                            count = videoItem.statistics?.viewCount ?: "0",
-                            dateTime = videoItem.snippet.publishedAt // API의 응답을 바탕으로 수정했습니다.
-                        )
-                    }
-
-                    // 변환된 HomeVideoModel 리스트를 어댑터에 적용
-                    homeVideoAdapter.setItems(videoModels)
-                } else {
-                    // 오류 처리
-                    Toast.makeText(requireContext(), "Error fetching videos", Toast.LENGTH_SHORT).show()
-                }
-            }
-
-            override fun onFailure(call: Call<VideosModelList>, t: Throwable) {
-                // 로딩 완료
-                binding.progressBar.visibility = View.GONE
-                // 네트워크 오류나 데이터 처리 오류 등의 실패 사유를 처리합니다.
-                Toast.makeText(requireContext(), "Error: ${t.message}", Toast.LENGTH_SHORT).show()
-            }
-        })
+        viewModel.fetchPopularVideos(categoryId)
     }
 }
